requirements.txt
streamlit
tensorflow
numpy
pillow
altair==4.2.2
